chore: bunch of stuff

+++ lib/as-bs.ts
@@ -292,7 +292,7 @@ export namespace bs {
       src = cacheOutput;
     }
     let dst = target;
-    if (changetype<OBJECT>(target - TOTAL_OVERHEAD).rtSize != len) {
+    if (usize(changetype<OBJECT>(target - TOTAL_OVERHEAD).rtSize) != len) {
       // @ts-expect-error: __renew is a runtime builtin
       dst = __renew(target, len);
     }
